Extracted client_options to a separate module
