+++ sdk/openai/inference/src/clients/azure.rs
@@ -2,22 +2,10 @@ use std::sync::Arc;
 
 use crate::auth::AzureKeyCredential;
 use crate::models::CreateChatCompletionsRequest;
+use crate::options::AzureOpenAIClientOptions;
 use crate::CreateChatCompletionsResponse;
-use azure_core::{
-    self, builders::ClientOptionsBuilder, AppendToUrlQuery, ClientOptions, HttpClient, Method,
-    Policy, Result,
-};
+use azure_core::{self, Method, Policy, Result};
 use azure_core::{Context, Url};
-
-// TODO: Implement using this instead
-// typespec_client_core::json_model!(CreateChatCompletionsResponse);
-
-// TODO: I was not  able to  find ClientOptions  as  a derive macros
-#[derive(Clone, Debug, Default)]
-pub struct AzureOpenAIClientOptions {
-    client_options: ClientOptions,
-    api_service_version: AzureServiceVersion,
-}
 
 pub struct AzureOpenAIClient<'a> {
     endpoint: Url,
@@ -41,7 +29,7 @@ impl AzureOpenAIClient<'_> {
         let options = client_options.unwrap_or_default();
         let per_call_policies: Vec<Arc<dyn Policy>> = key_credential.clone().into();
 
-        let pipeline = Self::new_pipeline(per_call_policies);
+        let pipeline = Self::new_pipeline(per_call_policies, options.client_options.clone());
 
         Ok(AzureOpenAIClient {
             endpoint,
@@ -51,10 +39,12 @@ impl AzureOpenAIClient<'_> {
         })
     }
 
-    fn new_pipeline(per_call_policies: Vec<Arc<dyn Policy>>) -> azure_core::Pipeline {
+    fn new_pipeline(
+        per_call_policies: Vec<Arc<dyn Policy>>,
+        options: azure_core::ClientOptions,
+    ) -> azure_core::Pipeline {
         let crate_name = option_env!("CARGO_PKG_NAME");
         let crate_version = option_env!("CARGO_PKG_VERSION");
-        let options = azure_core::ClientOptions::default();
         let per_retry_policies = Vec::new();
 
         azure_core::Pipeline::new(
@@ -96,70 +86,5 @@ impl AzureOpenAIClient<'_> {
             .send::<CreateChatCompletionsResponse>(&self.context, &mut request)
             .await?;
         response.into_body().json().await
-    }
-}
-
-impl AzureOpenAIClientOptions {
-    pub fn builder() -> builders::AzureOpenAIClientOptionsBuilder {
-        builders::AzureOpenAIClientOptionsBuilder::new()
-    }
-}
-
-pub mod builders {
-    use super::*;
-
-    #[derive(Clone, Debug, Default)]
-    pub struct AzureOpenAIClientOptionsBuilder {
-        options: AzureOpenAIClientOptions,
-    }
-
-    impl AzureOpenAIClientOptionsBuilder {
-        pub(super) fn new() -> Self {
-            Self::default()
-        }
-        pub fn with_api_version(mut self, api_service_version: AzureServiceVersion) -> Self {
-            self.options.api_service_version = api_service_version;
-            self
-        }
-
-        pub fn build(&self) -> AzureOpenAIClientOptions {
-            self.options.clone()
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum AzureServiceVersion {
-    V2023_09_01Preview,
-    V2023_12_01Preview,
-    V2024_07_01Preview,
-}
-
-impl Default for AzureServiceVersion {
-    fn default() -> AzureServiceVersion {
-        AzureServiceVersion::get_latest()
-    }
-}
-
-impl AzureServiceVersion {
-    pub fn get_latest() -> AzureServiceVersion {
-        AzureServiceVersion::V2024_07_01Preview
-    }
-}
-
-impl From<AzureServiceVersion> for String {
-    fn from(version: AzureServiceVersion) -> String {
-        let as_str = match version {
-            AzureServiceVersion::V2023_09_01Preview => "2023-09-01-preview",
-            AzureServiceVersion::V2023_12_01Preview => "2023-12-01-preview",
-            AzureServiceVersion::V2024_07_01Preview => "2024-07-01-preview",
-        };
-        return String::from(as_str);
-    }
-}
-
-impl ToString for AzureServiceVersion {
-    fn to_string(&self) -> String {
-        String::from(self.clone())
     }
 }
