+++ sdk/openai/inference/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod auth;
 mod clients;
 mod models;
+mod options;
 
 pub use clients::azure::*;
 pub use clients::non_azure::*;
 pub use models::*;
+pub use options::*;
